requirements.txt
streamlit
ultralytics
folium
streamlit-folium
gdown
pillow
torch
torchvision
